requirements.txt

appnope==0.1.0
asgiref==3.2.5
backcall==0.1.0
colorama==0.4.3
DateTime==4.3
decorator==4.4.2
dj-database-url==0.5.0
Django==3.0.6
django-cors-headers==3.2.1
django-extensions==2.2.9
django-filter==2.2.0
django-guardian==2.2.0
django-heroku==0.3.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
prompt-toolkit==3.0.4
psycopg2==2.8.5
psycopg2-binary==2.8.5
ptyprocess==0.6.0
Pygments==2.6.1
PyJWT==1.7.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
style==1.1.0
traitlets==4.3.3
update==0.0.1
wcwidth==0.1.8
whitenoise==5.1.0
zope.interface==5.1.0